move last bits of helpers out into snapenv
