+++ snap/snapenv/snapenv_test.go
@@ -17,7 +17,7 @@
  *
  */
 
-package helpers
+package snapenv
 
 import (
 	"testing"
