+++ systemd/systemd.go
@@ -36,9 +36,9 @@ import (
 
 	"github.com/ubuntu-core/snappy/arch"
 	"github.com/ubuntu-core/snappy/dirs"
-	"github.com/ubuntu-core/snappy/helpers"
 	"github.com/ubuntu-core/snappy/logger"
 	"github.com/ubuntu-core/snappy/osutil"
+	"github.com/ubuntu-core/snappy/snap/snapenv"
 )
 
 var (
@@ -417,10 +417,10 @@ WantedBy={{.ServiceSystemdTarget}}
 		restartCond,
 		desc.Type,
 	}
-	allVars := helpers.GetBasicSnapEnvVars(wrapperData)
-	allVars = append(allVars, helpers.GetUserSnapEnvVars(wrapperData)...)
-	allVars = append(allVars, helpers.GetDeprecatedBasicSnapEnvVars(wrapperData)...)
-	allVars = append(allVars, helpers.GetDeprecatedUserSnapEnvVars(wrapperData)...)
+	allVars := snapenv.GetBasicSnapEnvVars(wrapperData)
+	allVars = append(allVars, snapenv.GetUserSnapEnvVars(wrapperData)...)
+	allVars = append(allVars, snapenv.GetDeprecatedBasicSnapEnvVars(wrapperData)...)
+	allVars = append(allVars, snapenv.GetDeprecatedUserSnapEnvVars(wrapperData)...)
 	wrapperData.EnvVars = "\"" + strings.Join(allVars, "\" \"") + "\"" // allVars won't be empty
 
 	if err := t.Execute(&templateOut, wrapperData); err != nil {
